finish first module, and something of the third one

+++ backend/src/shared/domain/coords.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass
 class Coords:
-    lat : int
-    long : int
+    lat : float
+    long : float
